ref(models): Break up large file: `manager.py` (#27918)
