+++ mypy.ini
@@ -16,7 +16,7 @@ files = src/sentry/api/bases/external_actor.py,
         src/sentry/api/serializers/models/team.py,
         src/sentry/api/validators/external_actor.py,
         src/sentry/api/validators/notifications.py,
-        src/sentry/db/models/manager.py,
+        src/sentry/db/models/manager/**/*.py,
         src/sentry/features/**/*.py,
         src/sentry/grouping/strategies/base.py,
         src/sentry/grouping/strategies/message.py,
@@ -66,6 +66,8 @@ no_implicit_reexport=True
 follow_imports = skip
 
 
+[mypy-celery.*]
+ignore_missing_imports = True
 [mypy-confluent_kafka.*]
 ignore_missing_imports = True
 [mypy-google.*]
